rename classes directory to better align withtheir purposes
