+++ v3/src/viewmodels/ControlInformationViewModel.ts
