+++ v3/src/viewmodels/LayoutViewModel.tsx
